some refactor and enhancement
