+++ src/components/availability-panel/drag-drop-context/styles.module.scss
@@ -3,12 +3,25 @@
   justify-content: space-around;
   padding: 2rem;
   margin: 0 2rem 0 2rem;
-  justify-content: space-around;
   position: relative;
   margin-bottom: 2rem;
 }
+
+.emptyArray{
+  width: 32rem;
+  text-align: center;
+}
+
+.statusMessage {
+  display: flex;
+  justify-content: center;
+  padding: 1rem;
+  font-weight: bold;
+  color: #041484;
+}
+
 .divider {
-  border-left: 1px solid rgb(219, 214, 214);
+  border-left: 1px solid #D9D9D9;
   position: absolute;
   height: 100%;
 }
@@ -16,44 +29,74 @@
 .searchBoxContainer {
   display: flex;
   flex-direction: column;
-  span {
+
+  span{
+    height: 1.5rem;
+    width: 100%;
+    font-size: 20px;
+    line-height: 24px;
     color: #041187;
-    font-size: 1.5rem;
+    margin-bottom: 0.5rem;
+  }
+
+  .searchText {
     cursor: pointer;
   }
+
   input {
     height: 2.5rem;
+    width: 32rem;
+    background: #FFFFFF;
+    border: 2px solid #646464;
+    border-radius: 5px;
   }
 }
+
 .heading {
   height: 2.5rem;
+  width: 32rem;
   margin: 0.5rem 0;
+  border-radius: 5px;
   background-color: rgb(209, 209, 209);
 }
 
 .taskItem {
-  border: 1px solid rgb(219, 219, 219);
   padding: 0.8rem;
-  height: 3rem;
   width: 30rem;
   margin: 1rem 0;
   display: flex;
   align-items: center;
   color: #041187;
   font-size: large;
+  border: 2px solid #E0E0E0;
+  border-radius: 5px;
 }
+
+.idleMember{
+  display: flex;
+  flex-direction: column;
+  margin: 0.5rem 0;
+}
+
 .memberCard {
   display: flex;
+  border: 2px solid #E0E0E0;
+  padding: 0.8rem;
   align-items: center;
-}
-.memberCard {
+  flex-wrap: wrap;
+
   img {
     width: 3.5rem;
     height: 3.5rem;
     border-radius: 50%;
   }
+
   span {
-    padding-left: 0.5rem;
+    padding-left: 20px;
+    font-weight: bold;
+    font-size: 15px;
+    line-height: 37px;
+    color: #646464;
   }
 }
 
@@ -61,14 +104,23 @@
   .divider {
     display: none;
   }
+
   .flexContainer {
     flex-direction: column;
+    margin: 0;
   }
+
   .searchBoxContainer {
-    margin-top: 2.5rem;
+    input {
+      width: 100%;
+    }
   }
+
+  .heading{
+    width: 100%;
+  }
+
   .taskItem {
     width: 100%;
-    height: 100%;
   }
 }
